Other: Extract `closeDropdownOnBlur()` to own file.
